Refactor to library crate
* Move code in main.rs into example

* Add lib.rs

* Remove sqlx code that should be handled by rusie_queue users
